V0

+++ src/arch/armv8/interrupts.c
@@ -34,12 +34,12 @@ void interrupts_arch_init()
     interrupts_cpu_enable(platform.arch.gic.maintenance_id, true);
 }
 
-void interrupts_arch_ipi_send(uint64_t target_cpu, uint64_t ipi_id)
+void interrupts_arch_ipi_send(size_t target_cpu, size_t ipi_id)
 {
     if (ipi_id < GIC_MAX_SGIS) gic_send_sgi(target_cpu, ipi_id);
 }
 
-void interrupts_arch_enable(uint64_t int_id, bool en)
+void interrupts_arch_enable(size_t int_id, bool en)
 {
     gic_set_enable(int_id, en);
     gic_set_prio(int_id, 0x01);
@@ -50,28 +50,28 @@ void interrupts_arch_enable(uint64_t int_id, bool en)
     }
 }
 
-bool interrupts_arch_check(uint64_t int_id)
+bool interrupts_arch_check(size_t int_id)
 {
     return gic_get_pend(int_id);
 }
 
-inline bool interrupts_arch_conflict(bitmap_t interrupt_bitmap, uint64_t int_id)
+inline bool interrupts_arch_conflict(bitmap_t interrupt_bitmap, size_t int_id)
 {
     return (bitmap_get(interrupt_bitmap, int_id) && int_id > GIC_CPU_PRIV);
 }
 
-void interrupts_arch_clear(uint64_t int_id)
+void interrupts_arch_clear(size_t int_id)
 {
     gic_set_act(int_id, false);
     gic_set_pend(int_id, false);
 }
 
-void interrupts_arch_vm_assign(vm_t *vm, uint64_t id)
+void interrupts_arch_vm_assign(vm_t *vm, size_t id)
 {
     vgic_set_hw(vm, id);
 }
 
-void interrupts_arch_vm_inject(vm_t *vm, uint64_t id)
+void interrupts_arch_vm_inject(vm_t *vm, size_t id)
 {
     vgic_inject(&vm->arch.vgicd, id, cpu.vcpu->id);
 }
