+++ src/arch/armv8/iommu.c
@@ -18,7 +18,7 @@
 #include <iommu.h>
 #include <arch/smmuv2.h>
 
-int iommu_arch_init()
+long iommu_arch_init()
 {
     if(platform.arch.smmu.base){
         smmu_init();
@@ -36,7 +36,7 @@ static int32_t iommu_vm_arch_init_ctx(vm_t *vm)
         /* Set up ctx bank to vm address space in an available ctx. */
         ctx_id = smmu_alloc_ctxbnk();
         if (ctx_id >= 0) {
-            uint64_t rootpt;
+            size_t rootpt;
             mem_translate(&cpu.as, vm->as.pt.root, &rootpt);
             smmu_write_ctxbnk(ctx_id, (void *)rootpt, vm->id);
             vm->iommu.arch.ctx_id = ctx_id;
@@ -49,7 +49,7 @@ static int32_t iommu_vm_arch_init_ctx(vm_t *vm)
     return ctx_id;
 }
 
-static int iommu_vm_arch_add(vm_t *vm, uint16_t mask, uint16_t id)
+static long iommu_vm_arch_add(vm_t *vm, uint16_t mask, uint16_t id)
 {
     int32_t vm_ctx = iommu_vm_arch_init_ctx(vm);
     uint16_t glbl_mask = vm->iommu.arch.global_mask;
@@ -74,19 +74,19 @@ static int iommu_vm_arch_add(vm_t *vm, uint16_t mask, uint16_t id)
     return 0;
 }
 
-inline int iommu_arch_vm_add_device(vm_t *vm, int id)
+inline long iommu_arch_vm_add_device(vm_t *vm, int id)
 {
     return iommu_vm_arch_add(vm, 0, id);
 }
 
-int iommu_arch_vm_init(vm_t *vm, const vm_config_t *config)
+long iommu_arch_vm_init(vm_t *vm, const vm_config_t *config)
 {
     vm->iommu.arch.global_mask = 
         config->platform.arch.smmu.global_mask | platform.arch.smmu.global_mask;
     vm->iommu.arch.ctx_id = -1;
 
     /* This section relates only to arm's iommu so we parse it here. */
-    for (int i = 0; i < config->platform.arch.smmu.group_num; i++) {
+    for (size_t i = 0; i < config->platform.arch.smmu.group_num; i++) {
         /* Register each group. */
         const struct smmu_group *group =
             &config->platform.arch.smmu.smmu_groups[i];
