+++ src/core/inc/ipc.h
@@ -21,17 +21,17 @@
 #include <mem.h>
 
 typedef struct ipc {
-    uint64_t base;
+    size_t base;
     size_t size;
-    uint64_t shmem_id;
+    size_t shmem_id;
     size_t interrupt_num;
-    uint64_t *interrupts;
+    intptr_t *interrupts;
 } ipc_t;
 
 typedef struct vm_config vm_config_t;
 
-int64_t ipc_hypercall(uint64_t arg0, uint64_t arg1, uint64_t arg2);
+long ipc_hypercall(size_t arg0, size_t arg1, size_t arg2);
 void ipc_init(const vm_config_t* vm_config, bool vm_master);
-shmem_t* ipc_get_shmem(uint64_t shmem_id);
+shmem_t* ipc_get_shmem(size_t shmem_id);
 
 #endif /* IPC_H */
