+++ src/core/inc/vm.h
@@ -33,17 +33,17 @@
 #include <ipc.h>
 
 typedef struct vm {
-    uint64_t id;
+    size_t id;
 
     const vm_config_t* config;
 
     spinlock_t lock;
     cpu_synctoken_t sync;
-    uint64_t master;
+    size_t master;
 
     list_t vcpu_list;
-    uint64_t cpu_num;
-    uint64_t cpus;
+    size_t cpu_num;
+    size_t cpus;
 
     addr_space_t as;
 
@@ -66,7 +66,7 @@ typedef struct vcpu {
     struct arch_regs* regs;
     vcpu_arch_t arch;
 
-    size_t /*uint64_t*/ id;
+    size_t id;
     uint32_t phys_id;
     bool active;
 
@@ -78,30 +78,30 @@ typedef struct vcpu {
 extern vm_t vm;
 extern struct config* vm_config_ptr;
 
-void vm_init(vm_t* vm, const vm_config_t* config, bool master, uint64_t vm_id);
-void vm_start(vm_t* vm, uint64_t entry);
-vcpu_t* vm_get_vcpu(vm_t* vm, uint64_t vcpuid);
+void vm_init(vm_t* vm, const vm_config_t* config, bool master, size_t vm_id);
+void vm_start(vm_t* vm, size_t entry);
+vcpu_t* vm_get_vcpu(vm_t* vm, size_t vcpuid);
 void vm_emul_add_mem(vm_t* vm, emul_mem_t* emu);
 void vm_emul_add_reg(vm_t* vm, emul_reg_t* emu);
-emul_handler_t vm_emul_get_mem(vm_t* vm, uint64_t addr);
-emul_handler_t vm_emul_get_reg(vm_t* vm, uint64_t addr);
-void vcpu_init(vcpu_t* vcpu, vm_t* vm, uint64_t entry);
+emul_handler_t vm_emul_get_mem(vm_t* vm, size_t addr);
+emul_handler_t vm_emul_get_reg(vm_t* vm, size_t addr);
+void vcpu_init(vcpu_t* vcpu, vm_t* vm, size_t entry);
 void vm_msg_broadcast(vm_t* vm, cpu_msg_t* msg);
-uint64_t vm_translate_to_pcpu_mask(vm_t* vm, uint64_t mask, size_t len);
-uint64_t vm_translate_to_vcpu_mask(vm_t* vm, uint64_t mask, size_t len);
+size_t vm_translate_to_pcpu_mask(vm_t* vm, size_t mask, size_t len);
+size_t vm_translate_to_vcpu_mask(vm_t* vm, size_t mask, size_t len);
 
-static inline int64_t vm_translate_to_pcpuid(vm_t* vm, uint64_t vcpuid)
+static inline long vm_translate_to_pcpuid(vm_t* vm, size_t vcpuid)
 {
     return bitmap_find_nth((bitmap_t)&vm->cpus, sizeof(vm->cpus) * 8,
                            vcpuid + 1, 0, true);
 }
 
-static inline uint64_t vm_translate_to_vcpuid(vm_t* vm, uint64_t pcpuid)
+static inline size_t vm_translate_to_vcpuid(vm_t* vm, size_t pcpuid)
 {
     return bitmap_count((bitmap_t)&vm->cpus, 0, pcpuid, true);
 }
 
-static inline int vm_has_interrupt(vm_t* vm, int int_id)
+static inline size_t vm_has_interrupt(vm_t* vm, size_t int_id)
 {
     return bitmap_get(vm->interrupt_bitmap, int_id);
 }
@@ -111,11 +111,11 @@ static inline int vm_has_interrupt(vm_t* vm, int int_id)
 void vm_arch_init(vm_t* vm, const vm_config_t* config);
 void vcpu_arch_init(vcpu_t* vcpu, vm_t* vm);
 void vcpu_run(vcpu_t* vcpu);
-uint64_t vcpu_readreg(vcpu_t* vcpu, uint64_t reg);
-void vcpu_writereg(vcpu_t* vcpu, uint64_t reg, uint64_t val);
-uint64_t vcpu_readpc(vcpu_t* vcpu);
-void vcpu_writepc(vcpu_t* vcpu, uint64_t pc);
+size_t vcpu_readreg(vcpu_t* vcpu, size_t reg);
+void vcpu_writereg(vcpu_t* vcpu, size_t reg, size_t val);
+size_t vcpu_readpc(vcpu_t* vcpu);
+void vcpu_writepc(vcpu_t* vcpu, size_t pc);
 void vcpu_arch_run(vcpu_t* vcpu);
-void vcpu_arch_reset(vcpu_t* vcpu, uint64_t entry);
+void vcpu_arch_reset(vcpu_t* vcpu, size_t entry);
 
 #endif /* __VM_H__ */
