+++ src/core/objcache.c
@@ -48,7 +48,7 @@ static slab_t* slab_create(objcache_t* oc, enum AS_SEC sec)
             memset(slab, 0, PAGE_SIZE);
             slab->header.cache = oc;
             slab->header.objsize =
-                ALIGN((sizeof(node_t) + oc->osize), sizeof(uint64_t));
+                ALIGN((sizeof(node_t) + oc->osize), sizeof(size_t));
             slab->header.objnum =
                 (PAGE_SIZE - sizeof(((slab_t*)NULL)->header)) /
                 slab->header.objsize;
@@ -56,7 +56,7 @@ static slab_t* slab_create(objcache_t* oc, enum AS_SEC sec)
             list_t* obj_lst = &(slab->header.free);
             list_init(obj_lst);
             void* obj = &(slab->slab[sizeof(slab->header)]);
-            for (int i = 0; i < slab->header.objnum; i++) {
+            for (size_t i = 0; i < slab->header.objnum; i++) {
                 list_push(obj_lst, obj);
                 obj += slab->header.objsize;
             }
@@ -84,10 +84,10 @@ static bool slab_free(slab_t* slab, void* obj)
     void* obj_addr = obj - sizeof(node_t);
 
     if (slab != NULL) {
-        if (((((uint64_t)slab) & ~(PAGE_SIZE - 1)) ==
-             (((uint64_t)obj_addr) &
+        if (((((size_t)slab) & ~(PAGE_SIZE - 1)) ==
+             (((size_t)obj_addr) &
               ~(PAGE_SIZE - 1))) &&  // obj is part of slab
-            ((((((uint64_t)obj_addr) & (PAGE_SIZE - 1)) -
+            ((((((size_t)obj_addr) & (PAGE_SIZE - 1)) -
                sizeof(slab->header)) %
               slab->header.objsize) == 0) &&  // is aligned to object in slab
             (*((node_t*)(obj - sizeof(node_t))) ==
@@ -104,7 +104,7 @@ static bool slab_free(slab_t* slab, void* obj)
 }
 static slab_t* slab_get(void* obj)
 {
-    return (slab_t*)(((uint64_t)obj) & ~(PAGE_SIZE - 1));
+    return (slab_t*)(((size_t)obj) & ~(PAGE_SIZE - 1));
 }
 
 static bool slab_full(slab_t* slab)
